fix(otel): merge duplicate test_otel.py to fix pytest collection
PR #5815 placed tests in util/metrics/test/test_otel.py, but
util/test/test_otel.py already exists on master. Same module name
without __init__.py causes pytest import collision. Merge the new
tests into the existing file and remove the duplicate.
